+++ tests/component/sections/edit-dataset-terms/EditDatasetTerms.spec.tsx
@@ -3,7 +3,10 @@ import { EditDatasetTerms } from '@/sections/edit-dataset-terms/EditDatasetTerms
 import { DatasetProvider } from '@/sections/dataset/DatasetProvider'
 import { LicenseRepository } from '@/licenses/domain/repositories/LicenseRepository'
 import { DatasetRepository } from '@/dataset/domain/repositories/DatasetRepository'
-import { EditDatasetTermsHelper } from '@/sections/edit-dataset-terms/EditDatasetTermsHelper'
+import {
+  EditDatasetTermsHelper,
+  EditDatasetTermsTabKey
+} from '@/sections/edit-dataset-terms/EditDatasetTermsHelper'
 import { DatasetMother } from '@tests/component/dataset/domain/models/DatasetMother'
 import {
   TermsOfUseMother,
@@ -13,6 +16,7 @@ import { Dataset } from '@/dataset/domain/models/Dataset'
 import { License } from '@/licenses/domain/models/License'
 import { Guestbook } from '@/guestbooks/domain/models/Guestbook'
 import { GuestbookRepository } from '@/guestbooks/domain/repositories/GuestbookRepository'
+import { WithRepositories } from '@tests/component/WithRepositories'
 
 const licenseRepository: LicenseRepository = {} as LicenseRepository
 const datasetRepository: DatasetRepository = {} as DatasetRepository
@@ -89,11 +93,15 @@ describe('EditDatasetTerms', () => {
     datasetRepository.getByPersistentId = cy.stub().resolves(dataset)
     datasetRepository.getByPrivateUrlToken = cy.stub().resolves(dataset)
     return (
-      <DatasetProvider
-        searchParams={{ persistentId: 'some-persistent-id', version: 'some-version' }}
-        repository={datasetRepository}>
-        {component}
-      </DatasetProvider>
+      <WithRepositories
+        datasetRepository={datasetRepository}
+        guestbookRepository={guestbookRepository}>
+        <DatasetProvider
+          searchParams={{ persistentId: 'some-persistent-id', version: 'some-version' }}
+          repository={datasetRepository}>
+          {component}
+        </DatasetProvider>
+      </WithRepositories>
     )
   }
 
@@ -134,6 +142,24 @@ describe('EditDatasetTerms', () => {
     })
   })
 
+  it('renders NotFoundPage when dataset is missing', () => {
+    datasetRepository.getByPersistentId = cy.stub().resolves(undefined)
+    datasetRepository.getByPrivateUrlToken = cy.stub().resolves(undefined)
+
+    cy.customMount(
+      <DatasetProvider
+        searchParams={{ persistentId: 'some-persistent-id', version: 'some-version' }}
+        repository={datasetRepository}>
+        <EditDatasetTerms
+          defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.datasetTerms}
+          licenseRepository={licenseRepository}
+        />
+      </DatasetProvider>
+    )
+
+    cy.findByTestId('not-found-page').should('exist')
+  })
+
   describe('Tab Navigation', () => {
     it('renders all three tabs', () => {
       const dataset = DatasetMother.create({
@@ -146,8 +172,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.datasetTerms}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -169,8 +193,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.datasetTerms}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -203,8 +225,6 @@ describe('EditDatasetTerms', () => {
               EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.restrictedFilesTerms
             }
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -229,8 +249,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.datasetTerms}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -249,8 +267,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.datasetTerms}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -289,8 +305,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.datasetTerms}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -312,6 +326,71 @@ describe('EditDatasetTerms', () => {
         'true'
       )
     })
+
+    it('switches tabs without unsaved modal when active tab is guest book', () => {
+      const dataset = DatasetMother.create({
+        license: mockLicenses[0],
+        termsOfUse: TermsOfUseMother.withoutCustomTerms()
+      })
+
+      cy.customMount(
+        withProviders(
+          <EditDatasetTerms
+            defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.guestbook}
+            licenseRepository={licenseRepository}
+          />,
+          dataset
+        )
+      )
+
+      cy.findByRole('tab', { name: 'Dataset Terms' }).click()
+      cy.findByText('Unsaved Changes').should('not.exist')
+      cy.findByRole('tab', { name: 'Dataset Terms' }).should('have.attr', 'aria-selected', 'true')
+    })
+
+    it('switches tabs without unsaved modal when active tab key is unknown', () => {
+      const dataset = DatasetMother.create({
+        license: mockLicenses[0],
+        termsOfUse: TermsOfUseMother.withoutCustomTerms()
+      })
+
+      cy.customMount(
+        withProviders(
+          <EditDatasetTerms
+            // Force an invalid key to hit the default branch in getCurrentFormDirtyState
+            defaultActiveTabKey={'unknown-tab' as unknown as EditDatasetTermsTabKey}
+            licenseRepository={licenseRepository}
+          />,
+          dataset
+        )
+      )
+
+      cy.findByRole('tab', { name: 'Dataset Terms' }).click()
+      cy.findByText('Unsaved Changes').should('not.exist')
+      cy.findByRole('tab', { name: 'Dataset Terms' }).should('have.attr', 'aria-selected', 'true')
+    })
+
+    it('does nothing when selecting the currently active tab', () => {
+      const dataset = DatasetMother.create({
+        license: mockLicenses[0],
+        termsOfUse: TermsOfUseMother.withoutCustomTerms()
+      })
+
+      cy.customMount(
+        withProviders(
+          <EditDatasetTerms
+            defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.datasetTerms}
+            licenseRepository={licenseRepository}
+          />,
+          dataset
+        )
+      )
+
+      cy.get('select').select('CC BY 4.0')
+      cy.findByRole('tab', { name: 'Dataset Terms' }).click()
+      cy.findByText('Unsaved Changes').should('not.exist')
+      cy.findByRole('tab', { name: 'Dataset Terms' }).should('have.attr', 'aria-selected', 'true')
+    })
   })
 
   describe('Dataset Terms Tab Integration', () => {
@@ -325,8 +404,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.datasetTerms}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -347,8 +424,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.datasetTerms}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -378,8 +453,6 @@ describe('EditDatasetTerms', () => {
               EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.restrictedFilesTerms
             }
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -408,8 +481,6 @@ describe('EditDatasetTerms', () => {
               EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.restrictedFilesTerms
             }
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -453,8 +524,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.guestbook}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -478,8 +547,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.guestbook}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -501,8 +568,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.guestbook}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -526,11 +591,15 @@ describe('EditDatasetTerms', () => {
       datasetRepository.getByPersistentId = cy.stub().resolves(dataset)
       datasetRepository.getByPrivateUrlToken = cy.stub().resolves(dataset)
       return (
-        <DatasetProvider
-          searchParams={{ persistentId: 'some-persistent-id', version: 'some-version' }}
-          repository={datasetRepository}>
-          {component}
-        </DatasetProvider>
+        <WithRepositories
+          datasetRepository={datasetRepository}
+          guestbookRepository={guestbookRepository}>
+          <DatasetProvider
+            searchParams={{ persistentId: 'some-persistent-id', version: 'some-version' }}
+            repository={datasetRepository}>
+            {component}
+          </DatasetProvider>
+        </WithRepositories>
       )
     }
 
@@ -540,8 +609,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.datasetTerms}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           undefined
         )
@@ -559,8 +626,6 @@ describe('EditDatasetTerms', () => {
               EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.restrictedFilesTerms
             }
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -582,8 +647,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.guestbook}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -601,8 +664,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.guestbook}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -621,8 +682,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={'unknown-tab-key' as unknown as never}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -644,8 +703,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.datasetTerms}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -670,8 +727,6 @@ describe('EditDatasetTerms', () => {
           <EditDatasetTerms
             defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.datasetTerms}
             licenseRepository={licenseRepository}
-            datasetRepository={datasetRepository}
-            guestbookRepository={guestbookRepository}
           />,
           dataset
         )
@@ -687,11 +742,15 @@ describe('EditDatasetTerms Mobile View', () => {
     datasetRepository.getByPersistentId = cy.stub().resolves(dataset)
     datasetRepository.getByPrivateUrlToken = cy.stub().resolves(dataset)
     return (
-      <DatasetProvider
-        searchParams={{ persistentId: 'some-persistent-id', version: 'some-version' }}
-        repository={datasetRepository}>
-        {component}
-      </DatasetProvider>
+      <WithRepositories
+        datasetRepository={datasetRepository}
+        guestbookRepository={guestbookRepository}>
+        <DatasetProvider
+          searchParams={{ persistentId: 'some-persistent-id', version: 'some-version' }}
+          repository={datasetRepository}>
+          {component}
+        </DatasetProvider>
+      </WithRepositories>
     )
   }
 
@@ -707,8 +766,6 @@ describe('EditDatasetTerms Mobile View', () => {
         <EditDatasetTerms
           defaultActiveTabKey={EditDatasetTermsHelper.EDIT_DATASET_TERMS_TABS_KEYS.datasetTerms}
           licenseRepository={licenseRepository}
-          datasetRepository={datasetRepository}
-          guestbookRepository={guestbookRepository}
         />,
         dataset
       )